feat: moved auth hooks + context to shared package
