+++ packages/auth-context/src/hooks/useApiClient.tsx
@@ -1,4 +1,4 @@
 import { useContext } from "react";
-import { ApiContext } from "../context/ApiContext/ApiContext";
+import { ApiContext } from "../ApiContext/ApiContext";
 
 export const useApiClient = () => useContext(ApiContext);
