+++ webapp/package.json
@@ -33,6 +33,7 @@
   },
   "devDependencies": {
     "@easymotion/openapi": "workspace:../openapi",
+    "@easymotion/auth-context": "workspace:../packages/auth-context",
     "@eslint/js": "^9.15.0",
     "@tanstack/react-query-devtools": "^5.63.0",
     "@testing-library/react": "^16.1.0",
