+++ webapp/src/components/profile/ProfileSettings/PasswordUpdate.tsx
@@ -9,7 +9,7 @@ import {
   Typography,
 } from "@mui/material";
 import { useState } from "react";
-import { useApiClient } from "../../../hooks/useApiClient";
+import { useApiClient } from "@easymotion/auth-context";
 import { ensurePasswordConstraints } from "../../../data/validators";
 
 type Message = {
