+++ webapp/src/hooks/useProfile.tsx
@@ -1,6 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { useAuth } from "./useAuth";
-import { useApiClient } from "./useApiClient";
+import { useAuth } from "@easymotion/auth-context";
+import { useApiClient } from "@easymotion/auth-context";
 import { AuthUserDto, UpdateAuthUserDto } from "@easymotion/openapi";
 import { useSnack } from "./useSnack";
